[🔨refactor] 카테고리 관리 모달 입력창 컴포넌트로 분리
