+++ src/constants/category/index.ts
@@ -1,0 +1,3 @@
+export const IMPOSSIBLE_ADD = '* 이미 존재하는 카테고리는 추가 불가능합니다.';
+export const IMPOSSIBLE_DELETE = '* 메뉴가 포함된 카테고리는 삭제 불가능합니다.';
+export const SELECT_USING_OPTION = '메뉴 주문시 옵션 선택 창 사용';
